add a technician form done

+++ service/api/service_rest/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class Technician(models.Model):
     first_name = models.CharField(max_length=200)
     last_name = models.CharField(max_length=200)
-    employee_id = models.PositiveSmallIntegerField()
+    employee_id = models.PositiveBigIntegerField()
 
 
 class AutomobileVO(models.Model):
